requirements.txt
pymongo[srv] 
pymongo 
pandas
numpy
python-dotenv
certifi
openpyxl
streamlit
xlsxwriter
matplotlib
seaborn
